Updated HelloWorld.sol
Bug fix:

line88: uint toTransfer = balance; → It was missing the uint command.
line90: msg.sender.transfer(toTransfer); → Like you mentioned before, this is how it is supposed to be
The code is failing in the original case, if you try to withdraw all your funds from the contract. That is gonna wipe itself and that is not gonna return for the contract owner. In the previous line we defined balance = 0;, after there is no funds to send to the owner.

Please correct me, if I am wrong!